+++ src/main/java/com/fittura/domain/product/product/dto/response/ProductResDto.java
@@ -3,7 +3,7 @@ package com.fittura.domain.product.product.dto.response;
 import com.fittura.domain.product.product.constant.ProductStatus;
 import com.fittura.domain.product.product.constant.ProductType;
 import com.fittura.domain.product.product.entity.Product;
-import com.fittura.domain.product.sku.dto.responseDto.SkuResDto;
+import com.fittura.domain.product.sku.dto.response.SkuResDto;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.util.List;
