+++ src/main/java/com/fittura/domain/product/product/repository/ProductRepositoryImpl.java
@@ -4,8 +4,8 @@ import com.fittura.domain.product.product.constant.ProductStatus;
 import com.fittura.domain.product.product.dto.response.ProductResDto;
 import com.fittura.domain.product.product.dto.response.ProductWithStockResDto;
 import com.fittura.domain.product.sku.constant.SkuStatus;
-import com.fittura.domain.product.sku.dto.responseDto.SkuResDto;
-import com.fittura.domain.product.sku.dto.responseDto.SkuWithStockResDto;
+import com.fittura.domain.product.sku.dto.response.SkuResDto;
+import com.fittura.domain.product.sku.dto.response.SkuWithStockResDto;
 import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
