+++ src/main/java/com/fittura/domain/product/sku/dto/response/SkuWithStockResDto.java
@@ -1,21 +1,25 @@
-package com.fittura.domain.product.sku.dto.responseDto;
+package com.fittura.domain.product.sku.dto.response;
 
 import com.fittura.domain.product.sku.constant.SkuStatus;
 import com.fittura.domain.product.sku.entity.ProductSku;
 import io.swagger.v3.oas.annotations.media.Schema;
 
-@Schema(description = "SKU 응답 DTO")
-public record SkuResDto(
+@Schema(description = "SKU 응답 DTO (관리자용)")
+public record SkuWithStockResDto(
     Long id,
     Long price,
+    Integer stockQuantity,
+    Integer reservedQuantity,
     SkuStatus status,
     String color,
     String material
 ) {
-    public static SkuResDto from(ProductSku sku) {
-        return new SkuResDto(
+    public static SkuWithStockResDto from(ProductSku sku) {
+        return new SkuWithStockResDto(
             sku.getId(),
             sku.getPrice(),
+            sku.getStockQuantity(),
+            sku.getReservedQuantity(),
             sku.getStatus(),
             sku.getColor(),
             sku.getMaterial()
