+++ src/test/java/com/fittura/domain/product/product/service/ProductServiceTest.java
@@ -17,8 +17,8 @@ import com.fittura.domain.product.product.repository.ProductRepository;
 import com.fittura.domain.product.sku.constant.SkuStatus;
 import com.fittura.domain.product.sku.dto.request.CompositionCreateReqDto;
 import com.fittura.domain.product.sku.dto.request.SkuCreateReqDto;
-import com.fittura.domain.product.sku.dto.responseDto.SkuResDto;
-import com.fittura.domain.product.sku.dto.responseDto.SkuWithStockResDto;
+import com.fittura.domain.product.sku.dto.response.SkuResDto;
+import com.fittura.domain.product.sku.dto.response.SkuWithStockResDto;
 import com.fittura.global.exception.ServiceException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
